Create multiple pages
* Use a router to point at a different page depending on the URL
* Add a page for the manim-rs project with relevant links
* Separate out the file structure for different pages
