+++ Cargo.toml
@@ -7,7 +7,8 @@ edition = "2021"
 
 [dependencies]
 yew = { version = "0.21", features = ["csr"] }
-yew-bootstrap = "0.7"
+yew-router = { version = "0.18" }
+yew-bootstrap = "0.10"
 gloo-net = "0.2"
 serde = { version = "1.0", features = ["derive"] }
 wasm-bindgen-futures = "0.4"
