+++ src/pages/about.rs
@@ -1,0 +1,8 @@
+use yew::prelude::*;
+
+#[function_component(About)]
+pub fn about() -> Html {
+    html! {
+        <h1>{ "This is the about page!" }</h1>
+    }
+}
