+++ src/pages/home.rs
@@ -1,0 +1,8 @@
+use yew::prelude::*;
+
+#[function_component(Home)]
+pub fn home() -> Html {
+    html! {
+        <h1>{ "This is the homepage!" }</h1>
+    }
+}
