+++ src/pages/mod.rs
@@ -1,0 +1,6 @@
+mod about;
+mod home;
+pub mod projects;
+
+pub use about::*;
+pub use home::*;
