+++ src/pages/projects/mod.rs
@@ -1,0 +1,5 @@
+mod manim;
+mod yew_tutorial;
+
+pub use manim::*;
+pub use yew_tutorial::*;
